move user api and serializer to users module
