+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/users/auth/serializers.py
@@ -1,10 +1,11 @@
 # Third Party Stuff
-from django.contrib.auth import get_user_model, password_validation
+from django.contrib.auth import password_validation
 from rest_framework import serializers
 
 # {{ cookiecutter.project_name }} Stuff
 from {{cookiecutter.main_module}}.users import services as user_services
 from {{cookiecutter.main_module}}.users.models import UserManager
+from {{cookiecutter.main_module}}.users.serializers import UserSerializer
 
 from . import tokens
 
@@ -25,23 +26,11 @@ class RegisterSerializer(serializers.Serializer):
         return UserManager.normalize_email(value)
 
 
-class UserSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = get_user_model()
-        fields = ('id', 'first_name', 'last_name', 'email')
-
-
-class AuthUserSerializer(serializers.ModelSerializer):
-    name = serializers.SerializerMethodField()
+class AuthUserSerializer(UserSerializer):
     auth_token = serializers.SerializerMethodField()
 
-    class Meta:
-        model = get_user_model()
-        fields = ('id', 'name', 'auth_token', 'email')
-
-    def get_name(self, obj):
-        return obj.get_full_name()
+    class Meta(UserSerializer.Meta):
+        fields = UserSerializer.Meta.fields + ['auth_token']
 
     def get_auth_token(self, obj):
         return tokens.get_token_for_user(obj, "authentication")
