+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/users/serializers.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+from . import models
+
+
+class UserSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = models.User
+        fields = ['id', 'first_name', 'last_name', 'email']
